+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `AirplaneRightIconNext`: New icon.
+- Feature: `AirplaneIconNext`: Add `direction` prop, defaults to `up-left`.
 - Feature: `TagInputWithAutocomplete`: Add `showSuggestionsOnFocus` prop, defaults to false.
 
 ## [10.11.0] - 2022-05-23